requirements.txt
# Production dependencies
fast-langdetect>=0.3.2
fastmcp>=2.10.6
redis>=6.2.0
langfuse>=3.2.1
tenacity>=8.2.3
jsonschema>=4.23.0
aiolimiter>=1.1.0

# Optional provider SDKs (install as needed)
google-genai>=0.6.0
openai>=1.40.0
mistralai>=0.4.0
ollama>=0.3.0

# Development dependencies (uncomment for development)
# pytest>=7.0.0
# pytest-cov>=4.0.0
# black>=22.0.0
# flake8>=5.0.0
